+++ aim/web/ui/src/components/ChartPanel/ChartGrid/ChartGrid.tsx
@@ -19,6 +19,7 @@ function ChartGrid({
   chartProps,
   readOnly = false,
   syncHoverState,
+  resizeMode,
 }: IChartGridProps): React.FunctionComponentElement<React.ReactNode> {
   function getGridSize(dataLength: number, index: number): GridSize {
     return (
@@ -31,7 +32,12 @@ function ChartGrid({
         const Component = CHART_TYPES_CONFIG[chartType];
         const gridSize = getGridSize(data.length, index);
         return (
-          <Grid key={index} item className='ChartGrid' xs={gridSize}>
+          <Grid
+            key={`${index}-${resizeMode}`}
+            item
+            className='ChartGrid'
+            xs={gridSize}
+          >
             <Component
               ref={chartRefs[index]}
               nameKey={nameKey}
@@ -40,6 +46,7 @@ function ChartGrid({
               readOnly={readOnly}
               data={chartData}
               syncHoverState={syncHoverState}
+              resizeMode={resizeMode}
             />
           </Grid>
         );