+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fix the image's original size cropping (VkoHov)
 - Fix `PATH` related issues for `alembic` and `uvicorn` (mihran113)
 - Fix queries for custom object APIs (mihran113) 
+- Fix chart height updating when resize mode changed (VkoHov)
 
 ## 3.8.1 Apr 6, 2022